removing checks for auth

+++ Osa4/controllers/blogs.js
@@ -2,6 +2,7 @@ const blogsRouter = require('express').Router()
 const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const { requestLogger } = require('../utils/middleware')
 
 const getTokenFrom = request => {
       const authorization = request.get('authorization')
@@ -19,10 +20,14 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
     const token = request.token
-    const decodedToken = jwt.verify(token, process.env.SECRET)
-    if (!token || !decodedToken.id) {
-        return response.status(401).json({ error: 'token missing or invalid' })  
+    if (!token ) {
+        return response.status(401).json({ error: 'token missing or invalid' })
     }
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id){
+        return response.status(401).json({ error: 'token missing or invalid' })
+    } 
+    
 
     try{ 
         const user = await User.findById(decodedToken.id)
@@ -42,8 +47,19 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    if (!request.token) {
+        console.log('no token')
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
      try {
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+        const blog = await Blog.findById(request.params.id)
+        if (!decodedToken.id || decodedToken.id !== blog.user.toString()) {
+            console.log(decodedToken.id, blog.user.toString())
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
         await Blog.findByIdAndDelete(request.params.id)
+
         response.status(204).end()
          
      } catch (e) {
